fix(sources): add DNS-resolution SSRF guard to web_page reader
A hostname whose text is public can still resolve to a loopback, private,
link-local, or cloud-metadata address (169.254.169.254) at lookup time. Install
a reqwest DNS resolver that only yields globally routable addresses and
re-apply the host/scheme check on every redirect hop, so the fetch is pinned
to a vetted address. Extract the guard into web_page_ssrf.rs (with sibling
tests) to keep web_page.rs under the 500-line guideline. Enables the 'net'
tokio feature for the resolver's lookup_host call.
