+++ src/memory/sources/readers/web_page_tests.rs
@@ -155,102 +155,6 @@ fn tag_name_reads_leading_identifier() {
     assert_eq!(tag_name(">"), None);
 }
 
-// ── SSRF guard ──────────────────────────────────────────────────────
-
-#[test]
-fn is_url_allowed_accepts_public_http_urls() {
-    assert!(is_url_allowed(
-        &reqwest::Url::parse("https://example.com").unwrap()
-    ));
-    assert!(is_url_allowed(
-        &reqwest::Url::parse("http://example.com/x").unwrap()
-    ));
-    assert!(is_url_allowed(
-        &reqwest::Url::parse("https://sub.example.com").unwrap()
-    ));
-    assert!(is_url_allowed(
-        &reqwest::Url::parse("https://8.8.8.8").unwrap()
-    ));
-}
-
-#[test]
-fn is_url_allowed_rejects_private_and_internal_targets() {
-    // Private / loopback / link-local IP literals.
-    assert!(!is_url_allowed(
-        &reqwest::Url::parse("http://127.0.0.1").unwrap()
-    ));
-    assert!(!is_url_allowed(
-        &reqwest::Url::parse("http://10.0.0.1").unwrap()
-    ));
-    assert!(!is_url_allowed(
-        &reqwest::Url::parse("http://192.168.1.1").unwrap()
-    ));
-    assert!(!is_url_allowed(
-        &reqwest::Url::parse("http://169.254.169.254").unwrap()
-    ));
-    assert!(!is_url_allowed(
-        &reqwest::Url::parse("http://[::1]").unwrap()
-    ));
-    // Internal service names and local-only names.
-    assert!(!is_url_allowed(
-        &reqwest::Url::parse("http://localhost").unwrap()
-    ));
-    assert!(!is_url_allowed(
-        &reqwest::Url::parse("http://mongo").unwrap()
-    ));
-    assert!(!is_url_allowed(
-        &reqwest::Url::parse("http://service.internal").unwrap()
-    ));
-    // Non-http scheme.
-    assert!(!is_url_allowed(
-        &reqwest::Url::parse("ftp://example.com").unwrap()
-    ));
-    assert!(!is_url_allowed(
-        &reqwest::Url::parse("file:///etc/passwd").unwrap()
-    ));
-}
-
-#[test]
-fn is_blocked_host_rejects_ip_ranges_and_local_names() {
-    let blocked = [
-        "127.0.0.1",
-        "0.0.0.0",
-        "10.0.0.1",
-        "172.16.0.1",
-        "192.168.0.1",
-        "169.254.169.254",
-        "100.64.0.1", // CGNAT
-        "192.0.0.1",  // IETF protocol assignments
-        "localhost",
-        "foo.local",
-        "bar.internal",
-        "mongo",
-        "::1",
-        "fc00::1", // unique-local
-        "fe80::1", // link-local
-    ];
-    for host in blocked {
-        assert!(is_blocked_host(host), "expected {host:?} to be blocked");
-    }
-}
-
-#[test]
-fn is_blocked_host_accepts_public_hosts() {
-    let allowed = [
-        "8.8.8.8",
-        "1.1.1.1",
-        "example.com",
-        "sub.example.com",
-        "example.co.uk",
-        "8.8.8.8.",    // trailing dot is normalized away
-        "EXAMPLE.com", // case-insensitive
-        "2001:4860:4860::8888",
-    ];
-    for host in allowed {
-        assert!(!is_blocked_host(host), "expected {host:?} to be allowed");
-    }
-}
-
 // ── script/style stripping ──────────────────────────────────────────
 
 #[test]
